requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
pydantic==2.10.3
streamlit==1.41.1
requests==2.32.3

sentence-transformers==3.3.1
torch>=2.2.0

faiss-cpu==1.9.0.post1

pypdf==5.1.0
python-docx==1.1.2
numpy>=1.26.4,<3
httpx==0.27.2